# src/main/resources/static/api-docs.json
{"openapi":"3.0.1","info":{"title":"URLink App APIs","description":"By Dheeraj Krishna"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"URL APIs"}],"paths":{"/user/updatePassword":{"post":{"tags":["User APIs"],"summary":"Update current user's password","operationId":"updatePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/passwordDTO"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"*/*":{"schema":{"type":"boolean"}}}}},"security":[{"basicAuth":[]}]}},"/user/register":{"post":{"tags":["User APIs"],"summary":"Register a new user","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/userDTO"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/url":{"get":{"tags":["URL APIs"],"summary":"Get all added URLs","description":"Displays all the URLs of the user in a paginated format","operationId":"viewRedirects","parameters":[{"name":"pageNo","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":5}}],"responses":{"202":{"description":"Login successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UrlPage"}}}}},"security":[{"basicAuth":[]}]},"post":{"tags":["URL APIs"],"summary":"Add a new URL","description":"Add a new URL with a unique alias","operationId":"createRedirect","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/urlDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"basicAuth":[]}]},"delete":{"tags":["URL APIs"],"summary":"Delete a URL","description":"Delete a specific URL with it's alias","operationId":"DeleteRedirect","parameters":[{"name":"alias","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"basicAuth":[]}]}},"/url/update":{"post":{"tags":["URL APIs"],"summary":"Update URL","description":"Update the URL of an existing alias","operationId":"updateRedirect","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/urlUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"basicAuth":[]}]}},"/user/login":{"get":{"tags":["User APIs"],"summary":"Login the current user","description":"Logging in the current user using the basicAuth header","operationId":"loginUser","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/loginMessage"}}}}},"security":[{"basicAuth":[]}]}},"/url/{alias}":{"get":{"tags":["URL APIs"],"summary":"Redirect the user to the original URL","operationId":"HandleRedirect","parameters":[{"name":"alias","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"basicAuth":[]}]}},"/docs":{"get":{"tags":["Root APIs"],"summary":"Access documentation","description":"Access the redoc style documentation of the project","operationId":"getDocs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/":{"get":{"tags":["Root APIs"],"summary":"Root path","description":"Access the root path of the application","operationId":"homePage","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/user/{id}":{"delete":{"tags":["User APIs"],"summary":"Delete user with userId","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"basicAuth":[]}]}}},"components":{"schemas":{"passwordDTO":{"type":"object","properties":{"password":{"type":"string"}}},"userDTO":{"required":["firstName","password","role","userName"],"type":"object","properties":{"userName":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"},"role":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"Roles":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"authority":{"type":"string"}}},"Url":{"type":"object","properties":{"id":{"type":"integer","format":"int64","writeOnly":true},"alias":{"type":"string"},"url":{"type":"string"},"users":{"$ref":"#/components/schemas/User"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"},"password":{"type":"string","writeOnly":true},"roles":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Roles"}},"urls":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Url"}}}},"urlDTO":{"type":"object","properties":{"alias":{"type":"string"},"url":{"type":"string"}}},"urlUpdateDTO":{"type":"object","properties":{"alias":{"type":"string"},"newUrl":{"type":"string"}}},"loginMessage":{"type":"object","properties":{"message":{"type":"string"},"status":{"type":"boolean"},"user":{"$ref":"#/components/schemas/User"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}},"UrlPage":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"number":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/urlDTO"}},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}},"description":"Page of URLs"}},"securitySchemes":{"basicAuth":{"type":"http","name":"Authorization","in":"header","scheme":"basic"}}}}